requirements.txt
scrapy
python-dotenv
pytz
langdetect
python-dateutil